Inline LLM exchange details into trace event stream
Move LLM exchange expansion from a separate section at the bottom of
each call into the llm_exchange event rows themselves. Clicking an
exchange event header now lazy-loads and displays the full exchange
(system prompt, user message, response, tool calls) inline.

Also fix CORS to allow any localhost origin via regex instead of a
hardcoded port list.
